Changes - Update functions to trigger animations
I've made changes so that the players deck is shuffled to make things more tough.
The cards are now all flipped briefly for the player to remember.
Then players deck is shown.
As player selects cards they are flipped. If they select the incorrect card then they fail and all cards are flipped.
Cards are burned once player fails